break pafmask functions into their own source file
